anonymisation_service/requirements.txt
google-cloud-bigquery
logging
Jinja2
google-cloud-secret-manager
prefect
google-auth

